# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Illuminae API Documentation",
        "description": "",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://illuminate.io-devs.us/",
            "description": "Staging API Server"
        },
        {
            "url": "https://illuminate.io-devs.us/",
            "description": "Live API Server"
        },
        {
            "url": "http://backend_illuminate.test/",
            "description": "Local Server"
        },
        {
            "url": "http://192.168.0.143:8009/",
            "description": "Local Server Using IP"
        }
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "auth,register"
                ],
                "summary": "authentication",
                "description": "",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "full_name",
                        "in": "query",
                        "description": "Full Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_no",
                        "in": "query",
                        "description": "Phone Number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "login_type",
                        "in": "query",
                        "description": "Login Type ( 1=> Email, 2 => Google, 3 => Facebook, 4 => Apple ) ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "social_user_id",
                        "in": "query",
                        "description": "Social User Id ( required if login type is Google, Facebook or Apple )",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "social_token",
                        "in": "query",
                        "description": "Social Token ( required if login type is Google, Facebook )",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/verifyEmailVerificationCode": {
            "post": {
                "tags": [
                    "auth,verifyEmailVerificationCode"
                ],
                "summary": "authentication",
                "description": "",
                "operationId": "verifyEmailVerificationCode",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "verification_code",
                        "in": "query",
                        "description": "Verification Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/verificationCodeResend": {
            "post": {
                "tags": [
                    "auth,verificationCodeResend"
                ],
                "summary": "authentication",
                "description": "",
                "operationId": "verificationCodeResend",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/user/home/following": {
            "post": {
                "tags": [
                    "home,following"
                ],
                "summary": "following",
                "description": "",
                "operationId": "following",
                "parameters": [
                    {
                        "name": "follow_id",
                        "in": "query",
                        "description": "Follow user id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_follow",
                        "in": "query",
                        "description": "True or False",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}